[UPDATE]BUG

+++ blog-client/src/main/java/com/leyuna/blog/co/disk/UserFileInfoCO.java
@@ -17,7 +17,7 @@ public class UserFileInfoCO {
     /**
      * 用户内存总量
      */
-    private Double fileTotal;
+    private Long fileTotal;
 
     private String fileTotalStr;
